Fix null pointer errors in test code

+++ people-and-pets/src/test/java/au/com/cybersearch2/pp/PeopleAndPetsTest.java
@@ -165,13 +165,14 @@ public class PeopleAndPetsTest
     		
     	};
     	File[] files = defaultDatabaseFile.listFiles(filter);
-    	Arrays.asList(files).forEach(file -> {
-    		try {
-    			file.delete();
-    		} catch (Throwable t) {
-    			t.printStackTrace();
-    		}
-    	});
+    	if (files != null)
+	    	Arrays.asList(files).forEach(file -> {
+	    		try {
+	    			file.delete();
+	    		} catch (Throwable t) {
+	    			t.printStackTrace();
+	    		}
+	    	});
     }
 }
